# docs/admin/admin_swagger.yaml
basePath: /admin-api
definitions:
...
    handler.adminResponse:
        properties:
            code:
                type: integer
            data: {}
            msg:
                type: string
        type: object
    handler.adminRoleDocItem:
        properties:
            description:
                type: string
            enabled:
                type: boolean
            id:
                type: integer
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
            system:
                type: boolean
        type: object
...
    handler.createRoleReq:
        properties:
            description:
                type: string
            enabled:
                type: boolean
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
        required:
            - name
        type: object
...
    handler.permissionDocItem:
        properties:
            children:
                items:
                    $ref: '#/definitions/handler.permissionDocItem'
                type: array
            key:
                type: string
            label:
                type: string
        type: object
...
    handler.rolePermissionsResponse:
        properties:
            permissions:
                items:
                    type: string
                type: array
        type: object
    handler.updateRolePermReq:
        properties:
            permissions:
                items:
                    type: string
                type: array
        type: object
...
    model.AdminRole:
        properties:
            created_at:
                format: date-time
                type: string
            description:
                type: string
            enabled:
                type: boolean
            id:
                format: uint
                type: integer
            name:
                type: string
            permissions:
                items:
                    type: string
                type: array
            system:
                type: boolean
            updated_at:
                format: date-time
                type: string
        type: object
...
    swagger.PageResponse:
        properties:
            code:
                type: integer
            data:
                properties:
                    list:
                        items:
                            type: object
                        type: array
                    total:
                        format: int64
                        type: integer
                type: object
            msg:
                type: string
        type: object
    swagger.Response:
        properties:
            code:
                type: integer
            data:
                type: object
            msg:
                type: string
        type: object
host: localhost:8080
info:
    contact:
        name: API Support
        url: https://github.com/slowlyo/bico-admin
    description: 后台管理模块 API 文档
    license:
        name: MIT
        url: https://opensource.org/licenses/MIT
    termsOfService: https://github.com/slowlyo/bico-admin
    title: Bico Admin Admin API
    version: "1.0"
paths:
    /admin-roles:
        get:
            description: 需要权限：system:admin_role:list
            parameters:
                - description: 页码
                  in: query
                  name: page
                  required: false
                  type: integer
                - description: 每页数量
                  in: query
                  name: pageSize
                  required: false
                  type: integer
                - description: 排序字段
                  in: query
                  name: sortField
                  required: false
                  type: string
                - description: 排序方向：ascend 为升序，其余为降序
                  in: query
                  name: sortOrder
                  required: false
                  type: string
                - description: Name
                  in: query
                  name: name
                  required: false
                  type: string
                - description: Description
                  in: query
                  name: description
                  required: false
                  type: string
                - description: Enabled
                  in: query
                  name: enabled
                  required: false
                  type: boolean
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/swagger.PageResponse'
            security:
                - BearerAuth: []
            summary: 获取角色管理列表
            tags:
                - 角色管理
        post:
            description: 需要权限：system:admin_role:create
            parameters:
                - description: 创建参数
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.createRoleReq'
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/swagger.Response'
                            - properties:
                                data:
                                    $ref: '#/definitions/model.AdminRole'
                              type: object
            security:
                - BearerAuth: []
            summary: 创建角色管理
            tags:
                - 角色管理
...
    /admin-roles/{id}/permissions:
        get:
            description: 获取指定角色已配置的权限 key 列表
            parameters:
                - description: 角色 ID
                  in: path
                  name: id
                  required: true
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    $ref: '#/definitions/handler.rolePermissionsResponse'
                              type: object
            security:
                - BearerAuth: []
            summary: 获取角色权限
            tags:
                - 角色管理
        put:
            consumes:
                - application/json
            description: 覆盖指定角色的权限 key 列表
            parameters:
                - description: 角色 ID
                  in: path
                  name: id
                  required: true
                  type: integer
                - description: 权限列表
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.updateRolePermReq'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.adminResponse'
            security:
                - BearerAuth: []
            summary: 更新角色权限
            tags:
                - 角色管理
    /admin-roles/all:
        get:
            description: 获取下拉选择使用的启用角色列表
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    items:
                                        $ref: '#/definitions/handler.adminRoleDocItem'
                                    type: array
                              type: object
            security:
                - BearerAuth: []
            summary: 获取全部启用角色
            tags:
                - 角色管理
    /admin-roles/permissions:
        get:
            description: 获取后台所有菜单和按钮权限
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        allOf:
                            - $ref: '#/definitions/handler.adminResponse'
                            - properties:
                                data:
                                    items:
                                        $ref: '#/definitions/handler.permissionDocItem'
                                    type: array
                              type: object
            security:
                - BearerAuth: []
            summary: 获取完整权限树
            tags:
                - 角色管理
...
securityDefinitions:
    BearerAuth:
        description: 'JWT 认证，格式: Bearer {token}'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"